Moved all functions realate to unit commands and selection to map, reduced calculatePathToUnit to one function and fixed a bug, where the unit could walk through non-traversable tiles when the path was calculated by calculatePathToUnit
