+++ src/components/fragments/game/unit/Unit.js
@@ -1,5 +1,5 @@
 import PropTypes from "prop-types";
-import React, {useState} from "react";
+import React, {useEffect, useState} from "react";
 import UnitImage from "./UnitImage";
 import UnitShadow from "./UnitShadow";
 import UnitModel from "../../../../models/UnitModel";
@@ -9,8 +9,9 @@ import Keyframes from "../../../../helpers/Keyframes";
 
 const Unit = props => {
 
+    useEffect(() => {}, [props.unit.x, props.unit.y])
+
     const playerId = parseInt(localStorage.getItem("playerId"));
-    let playerIdCurrentTurn = parseInt(localStorage.getItem("playerIdCurrentTurn"));
 
     let tileSize = 48;
 
@@ -27,6 +28,10 @@ const Unit = props => {
         default:
             console.log("Team id " + props.unit.teamId + " not provided or does not match!");
             break;
+    }
+
+    const componentDidUpdate = (prevProps) => {
+        console.log(prevProps)
     }
 
     let unitStyle = {
@@ -64,7 +69,7 @@ const Unit = props => {
     }
 
     let markerStyle = {
-        animation: "unitMarkerMovement " + (1.0 + Math.random()) + "s infinite ease-in-out " + Math.random() +"s",
+        animation: "unitMarkerMovement 1.5s infinite ease-in-out",
     }
 
     return (
@@ -82,7 +87,7 @@ const Unit = props => {
             <UnitImage type={props.unit.type} color={unitColor} animation={animationState}/>
             <img src={healthSource} style={healthBoxStyle} alt={''}/>
             <div style={healthBarStyle} />
-            { (props.unit.userId === playerId && props.unit.performedAction === false && props.unit.userId === playerIdCurrentTurn) ?
+            { props.showMarker ?
                 <div className={"unitMarkerContainer"} style={markerContainerStyle}>
                     <img className={"unitMaker"} style={markerStyle} src={markerSource} alt={''} />
                 </div>
@@ -95,6 +100,7 @@ const Unit = props => {
 Unit.propTypes = {
     unit: PropTypes.instanceOf(UnitModel).isRequired,
     onClick: PropTypes.func.isRequired,
+    showMarker: PropTypes.bool,
 }
 
 export default Unit;
